refactored config reader
