+++ diktat-api/src/main/kotlin/com/saveourtool/diktat/Constants.kt
@@ -1,0 +1,16 @@
+/**
+ * This file contains common constants for Diktat
+ */
+package com.saveourtool.diktat
+
+
+/**
+ * Common application name, that is used in plugins and can be used to Suppress all diktat inspections on the
+ * particular code block with @Suppress("diktat")
+ */
+const val DIKTAT = "diktat"
+
+/**
+ * Default file name for config file
+ */
+const val DIKTAT_ANALYSIS_CONF = "diktat-analysis.yml"
