+++ diktat-api/src/main/kotlin/com/saveourtool/diktat/api/DiktatRuleNameAware.kt
@@ -1,0 +1,11 @@
+package com.saveourtool.diktat.api
+
+/**
+ * This interface represents *name* of individual inspection in rule set.
+ */
+interface DiktatRuleNameAware {
+    /**
+     * @return name of this [DiktatRule]
+     */
+    fun ruleName(): String
+}
