+++ diktat-cli/build.gradle.kts
@@ -23,10 +23,10 @@ dependencies {
     implementation(libs.log4j2.core)
     implementation(libs.log4j2.slf4j2)
 
-    testImplementation(projects.diktatConfiguration)
     testImplementation(projects.diktatKtlintEngine)
     testImplementation(projects.diktatRules)
     testImplementation(projects.diktatCommonTest)
+    testImplementation(libs.kaml)
     testImplementation(libs.junit.jupiter)
     testImplementation(libs.junit.platform.suite)
     testImplementation(libs.assertj.core)
