--- diktat-configuration/src/test/resources/log4j2.properties
@@ -1,8 +1,0 @@
-rootLogger.level = info
-rootLogger.appenderRef.stdout.ref = STDOUT
-
-appender.stdout.type = Console
-appender.stdout.name = STDOUT
-appender.stdout.target = SYSTEM_OUT
-appender.stdout.layout.type = PatternLayout
-appender.stdout.layout.pattern = [%-5p] %d{yyyy-MM-dd HH:mm:ss} %m%n
