+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/ruleset/config/AbstractDiktatRuleConfigReader.kt
@@ -1,8 +1,10 @@
-package com.saveourtool.diktat.config
+package com.saveourtool.diktat.ruleset.config
 
 import com.saveourtool.diktat.api.DiktatRuleConfig
 import com.saveourtool.diktat.api.DiktatRuleConfigReader
+import com.saveourtool.diktat.ruleset.constants.Warnings
 import io.github.oshai.kotlinlogging.KotlinLogging
+import org.jetbrains.kotlin.org.jline.utils.Levenshtein
 import java.io.IOException
 import java.io.InputStream
 import kotlin.jvm.Throws
@@ -19,7 +21,8 @@ abstract class AbstractDiktatRuleConfigReader : DiktatRuleConfigReader {
      * @param inputStream - input stream
      * @return list of [DiktatRuleConfig] if resource has been parsed successfully
      */
-    override fun invoke(inputStream: InputStream): List<DiktatRuleConfig> = read(inputStream).orEmpty()
+    override fun invoke(inputStream: InputStream): List<DiktatRuleConfig> =
+            read(inputStream)?.onEach(::validate).orEmpty()
 
     private fun read(inputStream: InputStream): List<DiktatRuleConfig>? = try {
         parse(inputStream)
@@ -39,6 +42,12 @@ abstract class AbstractDiktatRuleConfigReader : DiktatRuleConfigReader {
      */
     @Throws(IOException::class)
     protected abstract fun parse(inputStream: InputStream): List<DiktatRuleConfig>
+
+    private fun validate(config: DiktatRuleConfig) =
+        require(config.name == DIKTAT_COMMON || config.name in Warnings.names) {
+            val closestMatch = Warnings.names.minByOrNull { Levenshtein.distance(it, config.name) }
+            "Warning name <${config.name}> in configuration file is invalid, did you mean <$closestMatch>?"
+        }
 
     companion object {
         private val log = KotlinLogging.logger {}
