+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/ruleset/config/RuleConfiguration.kt
@@ -1,0 +1,7 @@
+package com.saveourtool.diktat.ruleset.config
+
+/**
+ * Configuration that allows customizing additional options of particular rules.
+ * @property config a map of strings with configuration options for a particular rule
+ */
+open class RuleConfiguration(protected val config: Map<String, String>)
