+++ diktat-rules/src/test/kotlin/com/saveourtool/diktat/ruleset/chapter2/KdocMethodsTest.kt
@@ -1,6 +1,6 @@
 package com.saveourtool.diktat.ruleset.chapter2
 
-import com.saveourtool.diktat.common.config.rules.DIKTAT_COMMON
+import com.saveourtool.diktat.ruleset.config.DIKTAT_COMMON
 import com.saveourtool.diktat.common.config.rules.DIKTAT_RULE_SET_ID
 import com.saveourtool.diktat.common.config.rules.RulesConfig
 import com.saveourtool.diktat.ruleset.constants.Warnings.KDOC_TRIVIAL_KDOC_ON_FUNCTION
