+++ diktat-rules/src/test/kotlin/com/saveourtool/diktat/ruleset/chapter3/TrailingCommaFixTest.kt
@@ -1,6 +1,6 @@
 package com.saveourtool.diktat.ruleset.chapter3
 
-import com.saveourtool.diktat.common.config.rules.DIKTAT_COMMON
+import com.saveourtool.diktat.ruleset.config.DIKTAT_COMMON
 import com.saveourtool.diktat.common.config.rules.RulesConfig
 import com.saveourtool.diktat.ruleset.constants.Warnings
 import com.saveourtool.diktat.ruleset.rules.chapter3.TrailingCommaRule
