+++ diktat-rules/src/test/resources/test-rules-config.yml
