+++ diktat-runner/src/main/kotlin/com/saveourtool/diktat/DiktatRunnerFactoryProvider.kt
@@ -7,7 +7,7 @@ package com.saveourtool.diktat
 import com.saveourtool.diktat.ktlint.DiktatBaselineFactoryImpl
 import com.saveourtool.diktat.ktlint.DiktatProcessorFactoryImpl
 import com.saveourtool.diktat.ktlint.DiktatReporterFactoryImpl
-import com.saveourtool.diktat.ruleset.rules.DiktatRuleConfigReaderImpl
+import com.saveourtool.diktat.ruleset.config.DiktatRuleConfigYamlReader
 import com.saveourtool.diktat.ruleset.rules.DiktatRuleSetFactoryImpl
 import generated.KTLINT_VERSION
 
@@ -20,7 +20,7 @@ const val ENGINE_INFO: String = "Ktlint: $KTLINT_VERSION"
  * @return initialized [DiktatRunnerFactory]
  */
 val diktatRunnerFactory: DiktatRunnerFactory = DiktatRunnerFactory(
-    DiktatRuleConfigReaderImpl(),
+    DiktatRuleConfigYamlReader(),
     DiktatRuleSetFactoryImpl(),
     DiktatProcessorFactoryImpl(),
     DiktatBaselineFactoryImpl(),
